backend/requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Scheduling
apscheduler>=3.10.4

# HTTP Client
httpx>=0.25.0

# Time Series Analysis (optional)
statsmodels>=0.14.0

# Configuration Management
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0

# Logging
loguru>=0.7.2

# Data Validation
pydantic-core>=2.10.0

# Additional utilities
openpyxl>=3.1.0
xlrd>=2.0.1